preprocessor: split preprocessor into its own file
