+++ cmake.ysfx.txt
@@ -55,6 +55,8 @@ add_library(ysfx-private
         "sources/ysfx_api_gfx_lice.hpp"
         "sources/ysfx_eel_utils.cpp"
         "sources/ysfx_eel_utils.hpp"
+        "sources/ysfx_preprocess.cpp"
+        "sources/ysfx_preprocess.hpp"
         "sources/utility/sync_bitset.hpp"
         "sources/base64/Base64.hpp")
 target_compile_definitions(ysfx-private
