+++ sources/ysfx_api_eel.cpp
@@ -47,7 +47,6 @@ static ysfx::mutex atomic_mutex;
 #include "WDL/eel2/eel_fft.h"
 #include "WDL/eel2/eel_mdct.h"
 #include "WDL/eel2/eel_atomic.h"
-#include "WDL/eel2/eel_pproc.h"
 
 //------------------------------------------------------------------------------
 void ysfx_api_init_eel()
@@ -145,30 +144,4 @@ void NSEEL_HOSTSTUB_EnterMutex()
 
 void NSEEL_HOSTSTUB_LeaveMutex()
 {
-}
-
-bool ysfx_preprocess(ysfx::text_reader &reader, ysfx_parse_error *error, std::string& in_str)
-{
-    std::string line;
-    uint32_t lineno = 0;
-
-    line.reserve(256);
-
-    WDL_FastString file_str, pp_str;
-    while (reader.read_next_line(line)) {
-        line += "\n";
-        const char *linep = line.c_str();
-        file_str.Append(linep);
-    }
-
-    EEL2_PreProcessor pproc;
-    const char *err = pproc.preprocess(file_str.Get(), &pp_str);
-    if (err) {
-        error->line = 0;
-        error->message = std::string("Invalid section: ") + err;
-        return false;
-    }
-
-    in_str.append(pp_str.Get(), pp_str.GetLength());
-    return true;
 }
